Backend/requirements.txt
flask
flask-cors
pdfplumber
pytesseract
Pillow
groq
python-dotenv